rename files - plus renamed amount to balance
